add statedict handler and improve joint model
